[Feat/#81] 공통 함수 익스텐션으로 처리
